requirements.txt
pyglet
pyglm
pyshaders
pycollada
